Rename the rule editor to library editor.
