+++ src/app/constants.js
@@ -14,5 +14,5 @@ export default {
   HOME: '/',
   LIB_SANDBOX: '/libSandbox',
   VIEW_SANDBOX: '/viewSandbox',
-  LIB_SANDBOX_RULES_EDITOR: '/libSandbox/rulesEditor'
+  LIB_SANDBOX_RULES_EDITOR: '/libSandbox/libraryEditor'
 };
