+++ src/app/libSandbox/libraryEditor/components/Backdrop.js
