+++ src/app/libSandbox/libraryEditor/components/ComponentIframe.css
