+++ src/app/libSandbox/libraryEditor/components/ExtensionConfigurationsList.js
