+++ src/app/libSandbox/libraryEditor/components/Settings.js
