+++ src/app/libSandbox/libraryEditor/helpers/basePath.js
@@ -10,4 +10,4 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-export default '/libSandbox/rulesEditor';
+export default '/libSandbox/libraryEditor';
