+++ src/app/libSandbox/libraryEditor/helpers/saveContainer.js
