+++ src/app/libSandbox/libraryEditor/models/currentRule.js
