+++ src/app/libSandbox/libraryEditor/models/otherSettings.js
