+++ src/app/libSandbox/libraryEditor/models/property.js
