+++ src/app/libSandbox/libraryEditor/models/registry.js
